Grabbed Someones code and fit it to our current

+++ pysite/views/api/bot/tags.py
@@ -1,3 +1,5 @@
+from contextlib import suppress
+
 from flask import jsonify
 from schema import Optional, Or, Schema
 
@@ -7,8 +9,9 @@ from pysite.decorators import api_key, api_params
 from pysite.mixins import DBMixin
 
 GET_SCHEMA = Schema({
-    Optional("tag_name"): str
+    Optional(Or("tag_name", "aliases_of")): str
 })
+
 
 POST_SCHEMA = Schema({
     "tag_name": str,
@@ -16,8 +19,15 @@ POST_SCHEMA = Schema({
     "image_url": Or(str, None, error="`image_url` must be none or a string")
 })
 
+
 DELETE_SCHEMA = Schema({
-    "tag_name": str
+    Or("tag_name", "aliases_of"): str
+})
+
+
+PATCH_SCHEMA = Schema({
+    "tag_name": str,
+    "tag_alias": str
 })
 
 
@@ -28,7 +38,7 @@ class TagsView(APIView, DBMixin):
 
     @api_key
     @api_params(schema=GET_SCHEMA, validation_type=ValidationTypes.params)
-    def get(self, params=None):
+    def get(self, params=None, aliases_of=None):
         """
         Fetches tags from the database.
 
@@ -45,13 +55,34 @@ class TagsView(APIView, DBMixin):
         API key must be provided as header.
         """
 
-        tag_name = None
+        if params is None:
+            tag_name = {}
 
-        if params:
-            tag_name = params.get("tag_name")
+        tag_name = params.get("tag_name")
+        aliases_of = params.get("aliases_of")
 
         if tag_name:
-            data = self.db.get(self.table_name, tag_name) or {}
+            data = (
+                self.db.get(self.table_name, tag_name)
+                or next(iter(self.db.filter(
+                        self.table_name,
+                        lambda row: row['tag_aliases'].contains(tag_name))), {})
+            )
+        elif aliases_of:
+            is_primary = self.db.get(self.table_name, aliases_of)
+            is_alias = self.db.filter(self.table_name, lambda row: row['tag_aliases'].contains(aliases_of))
+            if is_primary:
+                data = (
+                        [is_primary.get('tag_name')] + is_primary.get('tag_aliases')
+                )
+            elif is_alias:
+                document = next(iter(is_alias))
+                data = (
+                        [document.get('tag_name')] + document.get('tag_aliases')
+                )
+            else:
+                data = []
+
         else:
             data = self.db.pluck(self.table_name, "tag_name") or []
 
@@ -79,6 +110,7 @@ class TagsView(APIView, DBMixin):
             {
                 "tag_name": tag_name,
                 "tag_content": tag_content,
+                "tag_aliases": [],
                 "image_url": json_data.get("image_url")
             },
             conflict="update"  # If it exists, update it.
@@ -97,13 +129,43 @@ class TagsView(APIView, DBMixin):
         """
 
         tag_name = data.get("tag_name")
-        tag_exists = self.db.get(self.table_name, tag_name)
+        if tag_name:
+            tag_exists = self.db.get(self.table_name, tag_name)
+            if tag_exists:
+                self.db.delete(
+                    self.table_name,
+                    tag_name
+                )
+                return jsonify({"success": True})
+        alias_name = data.get('tag_alias')
+        if alias_name:
+            alias_exist = self.db.filter(self.table_name, lambda row: row['tag_aliases'].contains(alias_name))
+            if alias_exist:
+                self.db.update(
+                    self.table_name,
+                    lambda doc: doc['tag_aliases'].contains(alias_name),
+                    'tag_aliases',
+                    lambda db: db.row['tag_aliases'].set_difference([alias_name])
+                )
+                return jsonify({"success": True})
+        return jsonify({"success": False})
 
-        if tag_exists:
-            self.db.delete(
+    @api_key
+    @api_params(schema=PATCH_SCHEMA, validation_type=ValidationTypes.json)
+    def patch(self, data):
+        """
+        Adds an alias to database.
+
+        Data must be provided as JSON.
+        API key must be provided as header.
+        """
+        with suppress(ReferenceError):
+            self.db.update(
                 self.table_name,
-                tag_name
+                lambda doc: doc['tag_aliases'].contains(data['tag_name']) | (doc['tag_name'] == data['tag_name']),
+                'tag_aliases',
+                lambda db: db.row['tag_aliases'].default([]).append(data['tag_alias'])
             )
             return jsonify({"success": True})
 
-        return jsonify({"success": False})
+
